federator: Extract interpreter code to Federator.Interpreter
Delete unused runWaiError function and its tests. The tests were actually
testing AsWai instances of various errors, but they seem trivial enough to not
require tests.
